Fix warnings building with --std=c++11 and gcc 8
Silence warnings about signed constants being shifted out of int range.

+++ src/corelib/io/qdebug.h
@@ -96,7 +96,7 @@ class Q_CORE_EXPORT QDebug
         void setVerbosity(int v)
         {
             if (context.version > 1) {
-                flags &= ~(VerbosityMask << VerbosityShift);
+                flags &= ~(uint(VerbosityMask) << VerbosityShift);
                 flags |= (v & VerbosityMask) << VerbosityShift;
             }
         }
